move screenshot types to their local
